Added very basic UI at /BeerGame/GameView, added small console app to test logic, added a BeerGame javascript module this is used for any javascript we'll be programming.

+++ Blockchain Demonstrator Web App/Models/Classes/Farmer.cs
@@ -10,7 +10,7 @@ namespace Blockchain_Demonstrator_Web_App.Models.Classes
     public class Farmer : IRole
     {
         public Role Destination { get; set; } = Role.Processor;
-        public int LeadTime { get; set; } = 22;
+        public double LeadTime { get; set; } = 22;
         //TODO: replace dummy values
         public Dictionary<Option, IOption> Options { get; set; } = new Dictionary<Option, IOption>()
         {
